pybot/requirements.txt
python-dotenv
rich
azure-ai-inference
azure-core
